requirements.txt
django>=2.1.7
django-bootstrap4
django-environ
django-tinymce